Enum icu_calendar::any_calendar::AnyCalendar

#[non_exhaustive]
pub enum AnyCalendar {
    Gregorian(Gregorian),
    Buddhist(Buddhist),
    Japanese(Japanese),
    Indian(Indian),
    Coptic(Coptic),
    Iso(Iso),
}

This is a calendar that encompasses all formattable calendars supported by this crate

This allows for the construction of Date objects that have their calendar known at runtime.

This can be constructed by calling .into() on a concrete calendar type if the calendar type is known at compile time. When the type is known at runtime, the AnyCalendar::try_new_with_any_provider(), AnyCalendar::try_new_with_buffer_provider(), and AnyCalendar::try_new_unstable() methods may be used.

Date can also be converted to AnyCalendar-compatible ones via Date::to_any().

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Gregorian(Gregorian)

Buddhist(Buddhist)

Japanese(Japanese)

Indian(Indian)

Coptic(Coptic)

Iso(Iso)

Implementations

impl AnyCalendar

pub fn try_new_with_any_provider<P>(
kind: AnyCalendarKind,
provider: &P
) -> Result<Self, DataError> where
P: AnyProvider,

Constructs an AnyCalendar for a given calendar kind and AnyProvider data source

For calendars that need data, will attempt to load the appropriate data from the source

pub fn try_new_with_buffer_provider<P>(
kind: AnyCalendarKind,
provider: &P
) -> Result<Self, DataError> where
P: BufferProvider,

Constructs an AnyCalendar for a given calendar kind and BufferProvider data source

For calendars that need data, will attempt to load the appropriate data from the source

This needs the "serde" feature to be enabled to be used

pub fn try_new_unstable<P>(
kind: AnyCalendarKind,
provider: &P
) -> Result<Self, DataError> where
P: ResourceProvider<JapaneseErasV1Marker>,

Constructs an AnyCalendar for a given calendar kind and data source.

This method is unstable; the bounds on P might expand over time as more calendars are added

For calendars that need data, will attempt to load the appropriate data from the source

This needs the "serde" feature to be enabled to be used