Enum icu_calendar::any_calendar::AnyCalendarKind

#[non_exhaustive]
pub enum AnyCalendarKind {
    Gregorian,
    Buddhist,
    Japanese,
    JapaneseExtended,
    Ethiopian,
    EthiopianAmeteAlem,
    Indian,
    Coptic,
    Iso,
}

Convenient type for selecting the kind of AnyCalendar to construct

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Gregorian

The kind of a Gregorian calendar

Buddhist

The kind of a Buddhist calendar

Japanese

The kind of a Japanese calendar

JapaneseExtended

The kind of a JapaneseExtended calendar

Ethiopian

The kind of an Ethiopian calendar, with Amete Mihret era

EthiopianAmeteAlem

The kind of an Ethiopian calendar, with Amete Alem era

Indian

The kind of a Indian calendar

Coptic

The kind of a Coptic calendar

Iso

The kind of an Iso calendar

Implementations

impl AnyCalendarKind

pub fn get_for_bcp47_string(x: &str) -> Option<Self>

Construct from a BCP-47 string

Returns None if the calendar is unknown. If you prefer an error, use CalendarError::unknown_any_calendar_kind.

pub fn get_for_bcp47_bytes(x: &[u8]) -> Option<Self>

Construct from a BCP-47 byte string

Returns None if the calendar is unknown. If you prefer an error, use CalendarError::unknown_any_calendar_kind.

pub fn get_for_bcp47_value(x: &Value) -> Option<Self>

Construct from a BCP-47 Value

Returns None if the calendar is unknown. If you prefer an error, use CalendarError::unknown_any_calendar_kind.

pub fn as_bcp47_string(self) -> &'static str

Convert to a BCP-47 string

pub fn as_bcp47_value(self) -> Value

Convert to a BCP-47 Value

pub fn get_for_locale(l: &Locale) -> Option<Self>

Extract the calendar component from a Locale

Returns None if the calendar is not specified or unknown. If you prefer an error, use CalendarError::unknown_any_calendar_kind.