Trait icu_calendar::any_calendar::IncludedInAnyCalendar

pub trait IncludedInAnyCalendar: Calendar + Sized {
    fn to_any(self) -> AnyCalendar;
    fn to_any_cloned(&self) -> AnyCalendar;
    fn date_to_any(d: &Self::DateInner) -> AnyDateInner;
}

Trait for calendars that may be converted to AnyCalendar

Required Methods

fn to_any(self) -> AnyCalendar

Convert this calendar into an AnyCalendar, moving it

You should not need to call this method directly

fn to_any_cloned(&self) -> AnyCalendar

Convert this calendar into an AnyCalendar, cloning it

You should not need to call this method directly

fn date_to_any(d: &Self::DateInner) -> AnyDateInner

Convert a date for this calendar into an AnyDateInner

You should not need to call this method directly