icu_calendar::cal

Module east_asian_traditional_internal

Modules

china_data 🔒

Data obtained from calendrical_calculations.

korea_data 🔒

Data obtained from calendrical_calculations.

qing_data 🔒

This data is confirmed by both the Korea Astronomy and Space Science Institute, as well as the Hong Kong Observatory.

simple 🔒

Structs

China

The Rules used in China.

ChineseDateInner

The inner date type used for representing Dates of EastAsianTraditional.

EastAsianTraditional

The traditional East-Asian lunisolar calendar.

EastAsianTraditionalYearData

Information about a EastAsianTraditional year.

Korea

The Rules used in Korea.

PackedEastAsianTraditionalYearData 🔒

The struct containing compiled ChineseData

Traits

Rules

The rules for the EastAsianTraditional calendar.

Type Aliases

ChineseTraditional

The Chinese variant of the EastAsianTraditional calendar.

KoreanTraditional

The Korean variant of the EastAsianTraditional calendar.