icu_calendar::cal::east_asian_traditional_internal

Module simple

Macros

day_fraction_to_ms 🔒

Structs

LocalMoment 🔒

Milliseconds 🔒

Constants

BEIJING_UTC_OFFSET 🔒

MEAN_GREGORIAN_SOLAR_TERM_LENGTH 🔒

The mean solar term length according to the Gregorian solar cycle

MEAN_GREGORIAN_YEAR_LENGTH 🔒

The mean year length according to the Gregorian solar cycle.

MEAN_SYNODIC_MONTH_LENGTH 🔒

The mean synodic length on Jan 1 2000 according to the Astronomical Almanac (1992).

MILLISECONDS_IN_EPHEMERIS_DAY 🔒

Number of milliseconds in a day.

UTC_NEW_MOON 🔒

UTC_PLUS_8 🔒

UTC_PLUS_9 🔒

UTC_SOLSTICE 🔒