icu_calendar::cal::east_asian_traditional_internal

Struct China

#[non_exhaustive]
pub struct China;

The Rules used in China.

See ChineseTraditional for more information.

Implementations

impl China

pub fn gb_t_33661_2017(related_iso: i32) -> EastAsianTraditionalYearData

Computes EastAsianTraditionalYearData according to GB/T 33661-2017, as implemented by calendrical_calculations::chinese_based::Chinese.

The rules specified in GB/T 33661-2017 have only been used since 1912, applying them proleptically to years before 1912 will not necessarily match historical calendars.

Note that for future years there is a small degree of uncertainty, as GB/T 33661-2017 depends on the uncertain future difference between UT1 and UTC. As noted by Yuk Tung Liu, years as early as 2057, 2089, and 2097 have lunar events very close to local midnight, which might affect the start of a (single) month if additional leap seconds are introduced.