icu_calendar::cal

Module hijri_internal

Modules

simulated_mecca_data 🔒

Data obtained from calendrical_calculations.

ummalqura_data 🔒

Data obtained from ICU4C: https://github.com/unicode-org/icu/blob/1bf6bf774dbc8c6c2051963a81100ea1114b497f/icu4c/source/i18n/islamcal.cpp#L264

Structs

AstronomicalSimulation

Hijri Rules based on an astronomical simulation for a particular location.

Hijri

The Hijri Calendar

HijriDateInner

The inner date type used for representing Dates of Hijri. See Date and Hijri for more details.

HijriYearData

Information about a Hijri year.

PackedHijriYearData 🔒

The struct containing compiled Hijri YearInfo

TabularAlgorithm

Hijri Rules for the Tabular Hijri Algorithm.

UmmAlQura

Hijri Rules for the Umm al-Qura calendar.

Enums

SimulatedLocation 🔒

TabularAlgorithmEpoch

The epoch for the TabularAlgorithm rules.

TabularAlgorithmLeapYears

The leap year rule for the TabularAlgorithm rules.

Traits

Rules

Defines a variant of the Hijri calendar.