icu_calendar::cal

Struct JapaneseExtended

pub struct JapaneseExtended(Japanese);

The Japanese Calendar (with historical eras)

The Japanese Calendar is a variant of the Gregorian calendar created by the Japanese government. It is identical to the Gregorian calendar except that is uses Japanese eras instead of the Common Era.

This implementation extends proleptically for dates before the calendar’s creation in 6 Meiji (1873 CE). This implementation uses approximations of earlier Japanese eras proleptically and uses the Gregorian eras for even earlier dates that don’t have an approximate Japanese era.

For a variant whose Japanese eras start with Meiji, check out Japanese.

This corresponds to the "japanext" CLDR calendar.

Era codes

This calendar supports a large number of era codes. It supports the five eras since its introduction (meiji, taisho, showa, heisei, reiwa). Proleptic eras are represented with their names converted to lowercase ASCII and followed by their start year. E.g. the Ten’ō era (781 - 782 CE) has the code teno-781. The Gregorian bce (alias bc), and ce (alias ad) are used for dates before the first known era era.

These eras are loaded from data, requiring a data provider capable of providing CalendarJapaneseExtendedV1 data.

Tuple Fields
0: Japanese
Implementations

impl JapaneseExtended

pub const fn new() -> Self

Creates a new Japanese from using all eras (including pre-meiji) from compiled data.

✨ Enabled with the compiled_data Cargo feature.

📚 Help choosing a constructor

pub fn try_new_unstable<D: DataProvider<CalendarJapaneseExtendedV1> + ?Sized>( provider: &D, ) -> Result<Self, DataError>

A version of Self::new that uses custom data provided by a DataProvider.

📚 Help choosing a constructor

⚠️ The bounds on provider may change over time, including in SemVer minor releases.