icu_calendar::calendar_arithmetic

Trait PrecomputedDataSource

pub(crate) trait PrecomputedDataSource<YearInfo> {
    // Required method
    fn load_or_compute_info(&self, year: i32) -> YearInfo;
}

Required Methods

fn load_or_compute_info(&self, year: i32) -> YearInfo

Given a calendar year, load (or compute) the YearInfo for it

In the future we may pass in an optional previous YearInfo alongside the year it matches to allow code to take shortcuts.