icu_calendar::chinese_based

Trait ChineseBasedWithDataLoading

pub(crate) trait ChineseBasedWithDataLoading: Calendar {
    type CB: ChineseBased;

    // Required method
    fn get_precomputed_data(&self) -> ChineseBasedPrecomputedData<'_, Self::CB>;
}

The trait ChineseBased is used by Chinese-based calendars to perform computations shared by such calendar.

For an example of how to use this trait, see impl ChineseBasedWithDataLoading for Chinese in [Chinese].

Required Associated Types

type CB: ChineseBased

Required Methods

fn get_precomputed_data(&self) -> ChineseBasedPrecomputedData<'_, Self::CB>

Get the compiled const data for a ChineseBased calendar; can return None if the given year does not correspond to any compiled data.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.