icu_calendar

Module error

Error types for functions in the calendar crate

Modules

inner 🔒

unstable 🔒

Structs

RangeError

An argument is out of range for its domain.

UnknownEraError 🔒

Internal narrow error type for functions that only fail on unknown eras

Enums

DateError

Error type for date creation via Date::try_new_from_codes.

MonthCodeError 🔒

Internal narrow error type for functions that only fail on month code operations

MonthCodeParseError 🔒

Internal narrow error type for functions that only fail on parsing month codes

Functions

range_check 🔒

range_check_with_overflow 🔒