Struct icu_calendar::ethiopian::Ethiopian

pub struct Ethiopian(_);

The Ethiopian Calendar

The Ethiopian calendar is a solar calendar used by the Coptic Orthodox Church, with twelve normal months and a thirteenth small epagomenal month.

This type can be used with Date or DateTime to represent dates in this calendar.

It can be constructed in two modes: using the Amete Alem era scheme, or the Amete Mihret era scheme (the default), see EthiopianEraStyle for more info.

Era codes

This calendar supports three era codes, based on what mode it is in. In the Amete Mihret scheme it has the "incar" and "pre-incar" eras, 1 Incarnation is 9 CE. In the Amete Alem scheme, it instead has a single era, "mundi, where 1 Anno Mundi is 5493 BCE. Dates before that use negative year numbers.

Implementations

impl Ethiopian

pub fn new() -> Self

Construct a new Ethiopian Calendar for the Amete Mihret era naming scheme

pub fn new_with_era_style(era_style: EthiopianEraStyle) -> Self

Construct a new Ethiopian Calendar with a value specifying whether or not it is Amete Alem

pub fn set_era_style(&mut self, era_style: EthiopianEraStyle)

Set whether or not this uses the Amete Alem era scheme

pub fn era_style(&self) -> EthiopianEraStyle

Returns whether this has the Amete Alem era