Struct icu_calendar::indian::Indian

pub struct Indian;

The Indian National Calendar (aka the Saka calendar)

The Indian National calendar is a solar calendar used by the Indian government, with twelve months.

This type can be used with Date or DateTime to represent dates in this calendar.

Era codes

This calendar has a single era: "saka", with Saka 0 being 78 CE. Dates before this era use negative years.

Month codes

This calendar supports 12 solar month codes ("M01" - "M12")

Implementations

impl Indian

pub fn new() -> Self

Construct a new Indian Calendar