Struct icu_calendar::islamic::IslamicUmmAlQura

pub struct IslamicUmmAlQura { /* private fields */ }

Umm al-Qura Hijri Calendar (Used in Saudi Arabia)

Era codes

This calendar supports a single era code, Anno Mundi, with code "ah"

Month codes

This calendar is a pure lunar calendar with no leap months. It uses month codes "M01" - "M12".

Implementations

impl IslamicUmmAlQura

pub const fn new() -> Self

Creates a new IslamicUmmAlQura with some compiled data containing precomputed calendrical calculations.

โจ Enabled with the compiled_data Cargo feature.

๐ Help choosing a constructor

pub fn try_new_with_any_provider( provider: &(impl AnyProvider + ?Sized), ) -> Result<Self, CalendarError>

A version of Self::new that uses custom data provided by an AnyProvider.

๐ Help choosing a constructor

pub fn try_new_with_buffer_provider( provider: &(impl BufferProvider + ?Sized), ) -> Result<Self, CalendarError>

A version of Self::new that uses custom data provided by a BufferProvider.

โจ Enabled with the serde feature.

๐ Help choosing a constructor

pub fn try_new_unstable<D: DataProvider<IslamicUmmAlQuraCacheV1Marker> + ?Sized>( provider: &D, ) -> Result<Self, CalendarError>

A version of Self::new that uses custom data provided by a DataProvider.

๐ Help choosing a constructor

โ๏ธ The bounds on provider may change over time, including in SemVer minor releases.

pub fn new_always_calculating() -> Self

Construct a new IslamicUmmAlQura without any precomputed calendrical calculations.