icu_calendar::options

Module unstable

Structs

DateAddOptions

Options for adding a duration to a date.

DateDifferenceOptions

Options for taking the difference between two dates.

DateFromFieldsOptions

Options bag for Date::try_from_fields.

Enums

MissingFieldsStrategy

How to infer missing fields when the fields that are present do not fully constitute a Date.

Overflow

Whether to constrain or reject out-of-bounds values when constructing a Date.