icu_calendar::provider

Module hijri

🚧 [Unstable] Data provider struct definitions for chinese-based calendars.

🚧 This code is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. While the serde representation of data structs is guaranteed to be stable, their Rust representation might not be. Use with caution.

Read more about data providers: icu_provider

Structs

CalendarHijriSimulatedMeccaV1

Precomputed data for the Hijri obsevational calendar

HijriData

Cached/precompiled data for a certain range of years for a chinese-based calendar. Avoids the need to perform lunar calendar arithmetic for most calendrical operations.

PackedHijriYearInfo

The struct containing compiled Hijri YearInfo