icu_calendar::provider

Module islamic

🚧 [Unstable] Data provider struct definitions for chinese-based calendars.

🚧 This code is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. While the serde representation of data structs is guaranteed to be stable, their Rust representation might not be. Use with caution.

Read more about data providers: icu_provider

Structs

IslamicCacheV1

Cached/precompiled data for a certain range of years for a chinese-based calendar. Avoids the need to perform lunar calendar arithmetic for most calendrical operations.

IslamicObservationalCacheV1Marker

Marker type for IslamicCacheV1: “calendar/islamicobservationalcache@1”

IslamicUmmAlQuraCacheV1Marker

Marker type for IslamicCacheV1: “calendar/islamicummalquracache@1”

PackedIslamicYearInfo

The struct containing compiled Islamic YearInfo