Trait icu_calendar::Calendar

pub trait Calendar {
    type DateInner: PartialEq + Eq + Clone;
    fn date_from_iso(&self, iso: Date<Iso>) -> Self::DateInner;

    fn date_to_iso(&self, date: &Self::DateInner) -> Date<Iso>;

    fn months_in_year(&self, date: &Self::DateInner) -> u8;

    fn days_in_year(&self, date: &Self::DateInner) -> u32;

    fn days_in_month(&self, date: &Self::DateInner) -> u8;

    fn offset_date(
        &self, 
        date: &mut Self::DateInner, 
        offset: DateDuration<Self>
    );

    fn until(
        &self, 
        date1: &Self::DateInner, 
        date2: &Self::DateInner, 
        largest_unit: DurationUnit, 
        smallest_unit: DurationUnit
    ) -> DateDuration<Self>;

    fn debug_name() -> &'static str;

    fn day_of_week(&self, date: &Self::DateInner) -> u8 { ... }
}

A calendar implementation

Only implementors of Calendar should care about these methods, in general users of these calendars should use the methods on Date instead.

Individual Calendar implementations may have inherent utility methods allowing for direct construction, etc.

Associated Types

type DateInner: PartialEq + Eq + Clone

Loading content...
Required methods

fn date_from_iso(&self, iso: Date<Iso>) -> Self::DateInner

Construct the date from an ISO date

fn date_to_iso(&self, date: &Self::DateInner) -> Date<Iso>

fn months_in_year(&self, date: &Self::DateInner) -> u8

fn days_in_year(&self, date: &Self::DateInner) -> u32

fn days_in_month(&self, date: &Self::DateInner) -> u8

fn offset_date(&self, date: &mut Self::DateInner, offset: DateDuration<Self>)

Add offset to date

fn until(
&self,
date1: &Self::DateInner,
date2: &Self::DateInner,
largest_unit: DurationUnit,
smallest_unit: DurationUnit
) -> DateDuration<Self>

Calculate date2 - date as a duration

fn debug_name() -> &'static str

Loading content...
Provided methods

fn day_of_week(&self, date: &Self::DateInner) -> u8

Monday is 1, Sunday is 7, according to ISO

Loading content...