icu_calendar

Module types

This module contains various types used by icu_calendar and icu::datetime

Re-exports

pub use calendrical_calculations::rata_die::RataDie;

Structs

CyclicYear

Year information for a year that is specified as a cyclic year

DayOfMonth

A 1-based day number in a month.

DayOfWeekInMonth

A day of week in month. 1-based.

DayOfYear

The current day of the year, 1-based.

EraYear

Year information for a year that is specified with an era

IsoWeekOfYear

A week number in a year

MonthCode

Representation of a month in a year

MonthInfo

Representation of a formattable month.

Enums

Weekday

A weekday in a 7-day week, according to ISO-8601.

YearAmbiguity

Defines whether the era or century is required to interpret the year.

YearInfo

The type of year: Calendars like Chinese don’t have an era and instead format with cyclic years.