icu_calendar

Module week

Functions for region-specific weekday information.

Structs

UnitInfo 🔒

Information about a year or month.

WeekCalculator 🔒

WeekInformation

Information about the first day of the week and the weekend.

WeekOf 🔒

The week number assigned to a given week according to a calendar.

WeekPreferences

The preferences for the week information.

WeekdaySetIterator

Iterator that yields weekdays that are part of the weekend.

Enums

RelativeUnit 🔒

The year or month that a calendar assigns a week to relative to the year/month that it is in.

RelativeWeek 🔒

Which year or month that a calendar assigns a week to relative to the year/month the week is in.

Constants

MIN_UNIT_DAYS 🔒

Minimum number of days in a month unit required for using this module

Functions

add_to_weekday 🔒

Returns the weekday that’s num_days after weekday.