Module icu_calendar::any_calendar

Module for working with multiple calendars at once

Enums

AnyCalendar

This is a calendar that encompasses all formattable calendars supported by this crate

AnyCalendarKind

Convenient type for selecting the kind of AnyCalendar to construct

AnyDateInner

The inner date type for AnyCalendar

Traits

IntoAnyCalendar

Trait for calendars that may be converted to AnyCalendar