icu_calendar::cal

Struct JapaneseExtended

pub struct JapaneseExtended(/* private fields */);

The Japanese Calendar (with historical eras)

The Japanese calendar is a solar calendar used in Japan, with twelve months. The months and days are identical to that of the Gregorian calendar, however the years are counted differently using the Japanese era system.

This type can be used with Date to represent dates in this calendar.

Era codes

This calendar supports a large number of era codes. It supports the five post-Meiji eras (meiji, taisho, showa, heisei, reiwa). Pre-Meiji eras are represented with their names converted to lowercase ascii and followed by their start year. E.g. the Tenโล era (781 - 782 CE) has the code teno-781. The Gregorian bce (alias bc), and ce (alias ad) are used for dates before the first known era era.

These eras are loaded from data, requiring a data provider capable of providing CalendarJapaneseExtendedV1 data.

Month codes

This calendar supports 12 solar month codes (M01 - M12)

Implementations

impl JapaneseExtended

pub const fn new() -> Self

Creates a new Japanese from using all eras (including pre-meiji) from compiled data.

โจ Enabled with the compiled_data Cargo feature.

๐ Help choosing a constructor

pub fn try_new_with_buffer_provider( provider: &(impl BufferProvider + ?Sized), ) -> Result<Self, DataError>

A version of [Self :: new] that uses custom data provided by a BufferProvider.

โจ Enabled with the serde feature.

๐ Help choosing a constructor

pub fn try_new_unstable<D: DataProvider<CalendarJapaneseExtendedV1> + ?Sized>( provider: &D, ) -> Result<Self, DataError>

A version of Self::new that uses custom data provided by a DataProvider.

๐ Help choosing a constructor

โ๏ธ The bounds on provider may change over time, including in SemVer minor releases.