Trait icu_calendar::Calendar

pub trait Calendar {
    type DateInner: PartialEq + Eq + Clone;
    fn date_from_iso(&self, iso: Date<Iso>) -> Self::DateInner;

    fn date_to_iso(&self, date: &Self::DateInner) -> Date<Iso>;

    fn months_in_year(&self, date: &Self::DateInner) -> u8;

    fn days_in_year(&self, date: &Self::DateInner) -> u32;

    fn days_in_month(&self, date: &Self::DateInner) -> u8;

    fn offset_date(
        &self, 
        date: &mut Self::DateInner, 
        offset: DateDuration<Self>
    );

    fn until(
        &self, 
        date1: &Self::DateInner, 
        date2: &Self::DateInner, 
        largest_unit: DateDurationUnit, 
        smallest_unit: DateDurationUnit
    ) -> DateDuration<Self>;

    fn debug_name() -> &'static str;

    fn day_of_week(&self, date: &Self::DateInner) -> u8 { ... }
}

A calendar implementation

Only implementors of Calendar should care about these methods, in general users of these calendars should use the methods on Date instead.

Individual Calendar implementations may have inherent utility methods allowing for direct construction, etc.

Associated Types

type DateInner: PartialEq + Eq + Clone

The internal type used to represent dates

Required methods

fn date_from_iso(&self, iso: Date<Iso>) -> Self::DateInner

Construct the date from an ISO date

fn date_to_iso(&self, date: &Self::DateInner) -> Date<Iso>

Obtain an ISO date from this date

fn months_in_year(&self, date: &Self::DateInner) -> u8

Count the number of months in a given year, specified by providing a date from that year

fn days_in_year(&self, date: &Self::DateInner) -> u32

Count the number of days in a given year, specified by providing a date from that year

fn days_in_month(&self, date: &Self::DateInner) -> u8

Count the number of days in a given month, specified by providing a date from that year/month

fn offset_date(&self, date: &mut Self::DateInner, offset: DateDuration<Self>)

Add offset to date

fn until(
&self,
date1: &Self::DateInner,
date2: &Self::DateInner,
largest_unit: DateDurationUnit,
smallest_unit: DateDurationUnit
) -> DateDuration<Self>

Calculate date2 - date as a duration

fn debug_name() -> &'static str

Obtain a name for the calendar for debug printing

Provided methods

fn day_of_week(&self, date: &Self::DateInner) -> u8

Calculate the day of the week and return it

Monday is 1, Sunday is 7, according to ISO