icu_calendar

Module types

This module contains various types used by icu_calendar and icu::datetime

Macros

dt_unit 🔒

This macro defines a struct for 0-based date fields: hours, minutes, seconds and fractional seconds. Each unit is bounded by a range. The traits implemented here will return a Result on whether or not the unit is in range from the given input.

Structs

DayOfMonth

A day number in a month. Usually 1-based.

DayOfWeekInMonth

A day of week in month. 1-based.

DayOfYearInfo

A struct containing various details about the position of the day within a year. It is returned

Era

The era of a particular date

FormattableMonth

Representation of a formattable month.

FormattableYear

Representation of a formattable year.

IsoHour

An ISO-8601 hour component, for use with ISO calendars.

IsoMinute

An ISO-8601 minute component, for use with ISO calendars.

IsoSecond

An ISO-8601 second component, for use with ISO calendars.

MonthCode

Representation of a month in a year

NanoSecond

A fractional second component, stored as nanoseconds.

Time

A representation of a time in hours, minutes, seconds, and nanoseconds

WeekOfMonth

A week number in a month. Usually 1-based.

WeekOfYear

A week number in a year. Usually 1-based.

Enums

IsoWeekday

A weekday in a 7-day week, according to ISO-8601.