Module icu_calendar::week_of

Structs

UnitInfo 🔒

Information about a year or month.

WeekCalculator

Calculator for week-of-month and week-of-year based on locale-specific configurations.

WeekOf

The week number assigned to a given week according to a calendar.

WeekdaySetIterator

Iterator that yields weekdays that are part of the weekend.

Enums

RelativeUnit

The year or month that a calendar assigns a week to relative to the year/month that it is in.

RelativeWeek 🔒

Which year or month that a calendar assigns a week to relative to the year/month the week is in.

Constants

MIN_UNIT_DAYS

Minimum number of days in a month unit required for using this module

Functions

add_to_weekday 🔒

Returns the weekday that’s num_days after weekday.

simple_week_of

Computes & returns the week of given month or year according to a calendar with min_week_days = 1.

week_of

Computes & returns the week of given month/year according to calendar.